feat: Add flush_output functionality to main.F90
Ahora se actualiza (flush) la salida de forma forzada.
This commit adds the functionality to flush the output to a file in the main.F90 file. The `flush_output` procedure is implemented and called in the appropriate section of the code. This feature can be controlled by the `use_flush_output` logical parameter in the `parameters` module.