Code Upgrade
Now, instead of using intermediate steps to check for collisions/escapes,
there is a function called by the integrators callers that checks where to continue or not.
This ensures that, if a middle step (which could possibly be not taken into account by,
 for example, Bstoer) gives a collision or escape, it won't end up as a certain col/esc.
This new funtion takes away the job of hexitptr.