feat: Fix incorrect indexing in parameters.F90
This commit corrects the indexing issue in the parameters.F90 file. The code was incorrectly assigning values to the `rib` array, resulting in incorrect calculations for the center of mass of the asteroid and boulders. The indexing has been adjusted to assign the correct values to the `rib` array, ensuring accurate calculations for the center of mass.